+++ packages/admin/resources/views/partials/products/editing/sections.blade.php
@@ -109,6 +109,12 @@
         @include('adminhub::partials.products.editing.associations')
       </div>
 
+      {{--
+        Collections
+       --}}
+       <div id="collections">
+        @include('adminhub::partials.products.editing.collections')
+       </div>
 
       {{--
         Delete area