+++ packages/admin/src/Http/Livewire/Components/Products/ProductCreate.php
@@ -40,6 +40,7 @@ class ProductCreate extends AbstractProduct
 
         $this->syncAvailability();
         $this->syncAssociations();
+        $this->syncCollections();
     }
 
     /**